Don't wait for git mirror to be ready
Instead of waiting for the git repo to be cloned and ready, just start
the mirror going and be prepared for some operations to fail.

This is the first step towards operating without a git repo; the
second is to change (some) API responses depending on whether a git
repo is present, rather than failing when it's not.