Adapt minimally to the possible absence of a repo
Of the API methods, ListControllers is the only one for which a git
repo is optional; make it return a response with less information,
rather than an error, if the repo is not ready.

Some other methods either don't try to look at the repo (ListImages), in
which case no change is needed.

The remainder refer to the repo (usually because they want to write to
it), so must return an error. That will happen as things stand, since
attempting to clone the repo results in an error.